+++ src/main/javascript/Manifest/index.js
@@ -1,0 +1,11 @@
+const Builder = require('./Builder');
+const Manifest = require('./Manifest');
+const SyntaxValidator = require('./SyntaxValidator');
+const Resolver = require('./Resolver');
+
+module.exports = {
+  Builder: Builder,
+  Manifest: Manifest,
+  SyntaxValidator: SyntaxValidator,
+  Resolver: Resolver
+};