+++ src/main/javascript/Package/index.js
@@ -1,16 +1,9 @@
 'use strict';
 
 const path = require('path');
-const manifestSchema = require('../../resources/manifest.schema.json');
 const Validator = require('./Validator');
 
 exports = module.exports =
-  {
-    /**
-     * @returns {Validator}
-     */
-    newValidator: function () {
-      return new Validator({ manifestSchema: manifestSchema });
-    },
-
-  };
+{
+  Validator: Validator
+};